convert some interface[] returns to int64 for tests to pass

+++ problem13_test.go
@@ -9,7 +9,7 @@ import (
 func TestProblem13(t *testing.T) {
 	result := Problem13()
 
-	if result != 5537376230 {
+	if result != int64(5537376230) {
 		t.Error(errors.New(fmt.Sprintf("Problem13 failed. Expected 5537376230 Received %d", result)))
 	}
 }
